requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
requests==2.31.0
rsa==4.9
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
urlmatch==1.0.1
Werkzeug==3.0.2
